+++ interfaz/admin/barraNavegacionAdmin.php
@@ -5,11 +5,17 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Menú</title>
     <link rel="stylesheet" href="../../css/barraNavegacion.css">
+    <link rel='stylesheet' href='https://cdn-uicons.flaticon.com/2.4.2/uicons-solid-rounded/css/uicons-solid-rounded.css'>
     <script src="https://kit.fontawesome.com/7568cd4100.js" crossorigin="anonymous"></script>
 </head>
 <body id="body">
     <div class="menu__side" id="menu_side">
         <div class="options__menu">
+            <a href="usuarios.php" class="selected back">
+                <div class="option">
+                    <i class=" fi fi-sr-left" title="Usuarios"></i>
+                </div>
+            </a>
             <a href="pedidos.php?tipo=prep" class="selected">
                 <div class="option">
                     <i class="fa-solid fa-cubes" title="Pedidos"></i>
@@ -25,14 +31,14 @@
                     <i class="fa-solid fa-users" title="Usuarios"></i>
                 </div>
             </a>
-            <a href="transporte.php" class="selected">
+            <a href="abmTransportes.php" class="selected">
                 <div class="option">
                     <i class="fa-solid fa-truck-arrow-right" title="Transporte"></i>
                 </div>
             </a>
             <a href="reportes.php" class="selected">
                 <div class="option">
-                    <i class="fa-solid fa-book-open" title="Transporte"></i>
+                    <i class="fa-solid fa-book-open" title="Reportes Transporte"></i>
                 </div>
             </a>
             <a href="../../login.php" class="selected">