+++ js/admin/articulos.js
@@ -14,3 +14,4 @@ function mostrar(nro,texto){
   function modalEliminacion(idborrar){
     document.getElementById("id"+idborrar).style.display="block";
   }
+