+++ js/usuario/datosFinalesPedido.js
@@ -1,0 +1,7 @@
+function showSnackbar(mensaje) {
+  console.log("en tro");
+  var snackbar = document.getElementById("snackbar");
+  snackbar.innerText = mensaje;
+  snackbar.className = "show";
+  setTimeout(function() { snackbar.className = snackbar.className.replace("show", ""); }, 3000);
+}