+++ changelog.md
@@ -180,20 +180,49 @@ para continuar con el desarrollo del sistema.
     -/interfaz/admin/abmdestinatarios.php
     -/css/admin/abmdestinatarios.css
 
-## Añadida la opción para elegir transporte en Pedidos
+## Añadida la opción para elegir transporte en Pedidos  (Mauri)
 # Modificados:
     -/interfaz/admin/pedidos.php
     -/css/admin/pedidos.css
 
-## Creación interfaz para ABM Transportes 27/06/24
+## Creación interfaz para ABM Transportes 27/06/24 (Mauri)
 ## Nuevos:
     -/interfaz/admin/abmTransportes.php
     -/css/admin/abmTransportes.css
     -/js/admin/abmTransportes.js
 
-## Modificación estética en barra de navegación. 27/06/24 - Se agregó enlace a css externo para botón atrás y se corrigió un defecto por el cual no se veía correctamente el color. Se agregó clase "back" a dicho botón para luego recuperarlo en el js de cada pag. y configurar la página a la cual debe ir. Se agregaron 2 nuevos campos a mostrar en el apartado "proceso" de pedidos en Admin para que se muestre el codp y la localidad
+## Modificación estética en barra de navegación. 27/06/24 (Mauri) - Se agregó enlace a css externo para botón atrás y se corrigió un defecto por el cual no se veía correctamente el color. Se agregó clase "back" a dicho botón para luego recuperarlo en el js de cada pag. y configurar la página a la cual debe ir. Se agregaron 2 nuevos campos a mostrar en el apartado "proceso" de pedidos en Admin para que se muestre el codp y la localidad
 # Modificados:
     - /interfaz/admin/barranavegacion.php
     - /interfaz/admin/pedidos.php
     - /js/admin/pedidos.js
     - /interfaz/usuario/barranavegacion.php
+
+# Final pantalla de carga de destinatarios y datos finales 22/06/24 (diego)
+# modifiados:   
+    - css/login.css
+    - css/usuario/articulos.css
+    - css/usuario/datosFinalesPedido.css
+    - css/usuario/destinatarioPedido.css
+    - default.php
+    - interfaz/usuario/articulos.php
+    - interfaz/usuario/articulosDetalles.php
+    - interfaz/usuario/datosFinalesPedido.php
+    - interfaz/usuario/destinatarioPedido.php
+    - js/usuario/destinatarioPedido.js
+    - modelo/carrito.php
+    - modelo/conexion.php
+    - modelo/destinatario.php
+# nuevos:
+    - js/usuario/datosFinalesPedido.js
+
+# Final del proceso de creacion de pedido y mover los articulos de carrito a pedidos 23/06/2024 (diego)
+# modificados:
+    - css/usuario/datosFinalesPedido.css
+    - interfaz/usuario/datosFinalesPedido.php
+    - modelo/carrito.php
+# nuevos:
+    - css/usuario/pedidoProcesar.css
+    - interfaz/usuario/pedidoProcesar.php
+        
+## Merging nivelatorio con el repositorio 28/06/2024 (Mauri)